requirements.txt
asgiref==3.5.2
Django==4.0
django-environ==0.9.0
djangorestframework==3.14.0
install==1.3.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
